DAOS-8082 vos: Maintain a pool of dtx LID allocation arrays
Utilize the old IOF code to keep a pool of usable LRU tables
that can be used for containers when they have active DTX and
returned when not in use to avoid growing memory too much.

There are some improvements that we can likely make here at some
point.

1. Renewal may have potential optimization to avoid free/alloc and
perhaps only walk through a list of in-use slabs
2. If we have a good place to invoke d_slab_restock, we can avoid
doing allocations on the critical path.  For now, acquire will
do the allocation.

Allow-unstable-test: true

Signed-off-by: Jeff Olivier <jeffrey.v.olivier@intel.com>